Remove redundant parm to avoid incorrect usage
Both `detectInternalCycles` and `catchBlocksHaveRealPredecessors` are
intended to detect patterns in the CFG of the receiver method symbol.
When calling from `OMR::ResolvedMethodSymbol::genIL` the CFG of the
compiling method is passed in rather than the method whose IL is
generated which is incorrect.
Remove the CFG parm which can be obtained from self()->getFlowGraph() to
avoid incorrect use of the API.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>